requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
-e git+git://github.com/django-cms/django-cms.git@release/4.0.x#egg=django-cms
    # via
    #   -r requirements.in
    #   djangocms-alias
    #   djangocms-attributes-field
    #   djangocms-pageadmin
    #   djangocms-text-ckeditor
    #   djangocms-url-manager
    #   djangocms-version-locking
    #   djangocms-versioning
-e git+git://github.com/FidelityInternational/djangocms-pageadmin.git@0.0.14#egg=djangocms-pageadmin
    # via -r requirements.in
-e git+git://github.com/divio/djangocms-url-manager.git@0.0.11#egg=djangocms-url-manager
    # via -r requirements.in
-e git+git://github.com/FidelityInternational/djangocms-version-locking.git@0.0.14#egg=djangocms-version-locking
    # via
    #   -r requirements.in
    #   djangocms-pageadmin
boto3==1.17.106
    # via django-anymail
botocore==1.20.106
    # via
    #   boto3
    #   s3transfer
certifi==2021.5.30
    # via requests
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
click==8.0.1
    # via pip-tools
cryptography==3.3.2
    # via -r requirements.in
dj-database-url==0.5.0
    # via django-app-manage
django==2.2.24
    # via
    #   -r requirements.in
    #   django-anymail
    #   django-app-manage
    #   django-classy-tags
    #   django-cms
    #   django-debug-toolbar
    #   django-extensions
    #   django-filer
    #   django-formtools
    #   django-mptt
    #   django-polymorphic
    #   django-sekizai
    #   django-storages
    #   django-treebeard
    #   djangocms-alias
    #   djangocms-pageadmin
    #   djangocms-url-manager
    #   djangocms-version-locking
    #   djangocms-versioning
    #   easy-thumbnails
django-anymail[amazon_ses]==8.4
    # via -r requirements.in
https://github.com/ojii/django-app-manage/archive/65da18ef234a4e985710c2c0ec760023695b40fe.zip#egg=django-app-manage
    # via -r requirements.in
django-apptemplates==1.5
    # via -r requirements.in
django-classy-tags==2.0.0
    # via
    #   django-cms
    #   django-sekizai
django-debug-toolbar==3.2.1
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-filer==2.0.2
    # via -r requirements.in
django-formtools==2.3
    # via django-cms
django-fsm==2.6.1
    # via djangocms-versioning
django-js-asset==1.2.2
    # via django-mptt
django-mptt==0.12.0
    # via django-filer
django-parler==2.2
    # via djangocms-alias
django-polymorphic==3.0.0
    # via django-filer
django-sekizai==2.0.0
    # via django-cms
django-storages==1.11.1
    # via -r requirements.in
django-treebeard==4.5.1
    # via
    #   django-cms
    #   djangocms-pageadmin
djangocms-admin-style==2.0.2
    # via django-cms
https://github.com/divio/djangocms-alias/tarball/master#egg=djangocms-alias
    # via -r requirements.in
djangocms-attributes-field==1.2.0
    # via djangocms-url-manager
https://github.com/divio/djangocms-text-ckeditor/archive/support/4.0.x.zip
    # via -r requirements.in
https://github.com/divio/djangocms-versioning/tarball/master#egg=djangocms-versioning
    # via
    #   -r requirements.in
    #   djangocms-pageadmin
easy-thumbnails==2.7.1
    # via
    #   -r requirements.in
    #   django-filer
html5lib==1.1
    # via djangocms-text-ckeditor
idna==2.10
    # via requests
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
pep517==0.10.0
    # via pip-tools
pillow==8.3.1
    # via
    #   djangocms-text-ckeditor
    #   easy-thumbnails
pip-tools==6.2.0
    # via -r requirements.in
psycopg2==2.8.6
    # via -r requirements.in
pycparser==2.20
    # via cffi
python-dateutil==2.8.1
    # via botocore
python-dotenv==0.18.0
    # via -r requirements.in
pytz==2021.1
    # via
    #   -r requirements.in
    #   django
requests==2.25.1
    # via django-anymail
s3transfer==0.4.2
    # via boto3
six==1.16.0
    # via
    #   cryptography
    #   html5lib
    #   python-dateutil
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
toml==0.10.2
    # via pep517
unidecode==1.1.2
    # via django-filer
urllib3==1.26.6
    # via
    #   botocore
    #   requests
webencodings==0.5.1
    # via html5lib
werkzeug==2.0.1
    # via -r requirements.in
wheel==0.36.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
